backend/requirements.txt
fastapi
uvicorn[standard]
python-docx

python-multipart
